feat: 收藏添加标题, 添加搜索引擎搜索, webview默认滚动速度优化

+++ lib/app/widgets/star_dialog.dart
@@ -1,12 +1,15 @@
 import 'package:core/core.dart';
 import 'package:flicker_free/app/common/global.dart';
+import 'package:flicker_free/app/widgets/tag.dart';
 import 'package:flicker_free/db/entity/star.dart';
 import 'package:flicker_free/db/services/star_service.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 class StarDialog extends StatefulWidget {
-  const StarDialog({super.key});
+  final Star current;
+
+  const StarDialog({super.key, required this.current});
 
   @override
   State<StarDialog> createState() => _StarDialogState();
@@ -44,15 +47,28 @@ class _StarDialogState extends State<StarDialog> with AppMessageMixin {
             itemBuilder: (BuildContext context, int index) {
               var star = stars[index];
               return ListTile(
-                leading: Text('${star.id}'),
+                leading: Tag('${star.id}'),
                 title: Text(
+                  star.title ?? '-',
+                  maxLines: 1,
+                  overflow: TextOverflow.ellipsis,
+                  style: const TextStyle(fontWeight: FontWeight.bold),
+                ),
+                subtitle: Text(
                   '${star.url}',
                   maxLines: 1,
                   overflow: TextOverflow.ellipsis,
+                  style: const TextStyle(fontSize: 12),
                 ),
                 trailing: IconButton(
                   icon: const Icon(Icons.delete),
                   onPressed: () async {
+                    var id = widget.current.id;
+                    if (id == star.id) {
+                      await showToast('当前收藏被占用');
+                      return;
+                    }
+
                     await starService.remove(star.id);
                     await loadData();
                     await showToast('删除成功');
@@ -66,6 +82,9 @@ class _StarDialogState extends State<StarDialog> with AppMessageMixin {
           ),
         ),
       ),
+      actions: [
+        TextButton(onPressed: () => Get.back(), child: const Text('关闭')),
+      ],
     );
   }
 }
